+++ src/main/java/yesman/epicfight/client/particle/CustomModelParticle.java
@@ -62,9 +62,9 @@ public abstract class CustomModelParticle<M extends Mesh<?>> extends Particle {
 		rotation.mul(Vector3f.ZP.rotationDegrees(roll));
 		
 		Vec3 vec3 = camera.getPosition();
-		float x = (float)(Mth.lerp((double)partialTicks, this.xo, this.x) - vec3.x());
-		float y = (float)(Mth.lerp((double)partialTicks, this.yo, this.y) - vec3.y());
-		float z = (float)(Mth.lerp((double)partialTicks, this.zo, this.z) - vec3.z());
+		float x = (float)(Mth.lerp(partialTicks, this.xo, this.x) - vec3.x());
+		float y = (float)(Mth.lerp(partialTicks, this.yo, this.y) - vec3.y());
+		float z = (float)(Mth.lerp(partialTicks, this.zo, this.z) - vec3.z());
 		float scale = (float)Mth.lerp((double)partialTicks, this.scaleO, this.scale);
 		
 		poseStack.translate(x, y, z);
