+++ src/main/java/yesman/epicfight/world/entity/ai/attribute/EpicFightAttributes.java
@@ -110,7 +110,7 @@ public class EpicFightAttributes {
 	}
 
 	public static AttributeModifier getMaxStrikesModifier(int value) {
-		return new AttributeModifier(EpicFightAttributes.MAX_STRIKE_MODIFIER, EpicFightMod.MODID + ":weapon_modifier", (double) value, AttributeModifier.Operation.ADDITION);
+		return new AttributeModifier(EpicFightAttributes.MAX_STRIKE_MODIFIER, EpicFightMod.MODID + ":weapon_modifier", value, AttributeModifier.Operation.ADDITION);
 	}
 
 	public static AttributeModifier getImpactModifier(double value) {
